+++ src/c_cmds.c
@@ -7009,7 +7009,9 @@ static void spawn_cmd_func2(char *cmd, char *parms)
                 spawn = false;
             }
 
-            if (gamemode == shareware && (spawncmdtype == Cyberdemon || spawncmdtype == SpiderMastermind || spawncmdtype == Berserk))
+            if (gamemode == shareware
+                && (spawncmdtype == Cyberdemon || spawncmdtype == SpiderMastermind
+                    || spawncmdtype == Berserk || spawncmdtype == HelperDog))
             {
                 M_StringCopy(buffer, mobjinfo[type].plural1, sizeof(buffer));